refactor: fireworks
